productsearch: make bleve key value filters disjunctive (#3)
Previously when providing a key values filter all the values needed to be there. Example I provide a filter with brandCode = ["brandA", "brandB"] this means a product needs both brands which is rather useless. Therefore, changing this to a disjunctive query will now return products with either brandA or brandB set.